Move CDC code to separate python file
